requirements.txt
appnope==0.1.2
appscript==1.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.9
django-allauth==0.46.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
entrypoints==0.3
filelock==3.4.0
HeapDict==1.0.1
idna==3.3
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
oauthlib==3.1.1
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.6.0
zope.event==4.5.0